#909 refactor(agent): make Host own legacy route lifecycle
