+++ packages/agent/src/server/index.ts
@@ -144,6 +144,9 @@ export type {
   AgentHostHandle,
   AgentHostHttpProjectionOptions,
   AgentHostLegacyHttpProjectionOptions,
+  AgentHostLegacyProjectionComposition,
+  AgentHostLegacyProjectionLifecycle,
+  AgentHostLegacyProjectionRuntime,
   AgentHostLegacyRoutePolicy,
   AgentHostLegacyRoutePolicyMountInput,
   AgentRequestKey,
